game_test.go
package poker_test

import (
	poker "Go-Simple-Web-Server"
	"testing"
	"time"

	"github.com/stretchr/testify/assert"
)

func TestGame_Start(t *testing.T) {
	t.Run("schedules alerts on game starts for 5 players", func(t *testing.T) {
		blindAlerter := &SpyBlindAlerter{}
		game := poker.NewTexasHoldem(blindAlerter, dummyPlayerStore)

		game.Start(5)

		cases := []ScheduledAlert{
			{At: 0 * time.Second, Amount: 100},
			{At: 10 * time.Minute, Amount: 200},
			{At: 20 * time.Minute, Amount: 300},
			{At: 30 * time.Minute, Amount: 400},
			{At: 40 * time.Minute, Amount: 500},
			{At: 50 * time.Minute, Amount: 600},
			{At: 60 * time.Minute, Amount: 800},
			{At: 70 * time.Minute, Amount: 1000},
			{At: 80 * time.Minute, Amount: 2000},
			{At: 90 * time.Minute, Amount: 4000},
			{At: 100 * time.Minute, Amount: 8000},
		}
		assert.Equal(t, cases, blindAlerter.alerts)
	})
}

func TestGame_Finish(t *testing.T) {
	store := &poker.StubPlayerStore{}
	game := poker.NewTexasHoldem(dummySpyAlerter, store)

	winner := "George"

	game.Finish(winner)

	assert.Equal(t, winner, store.WinCalls[0])
}
